Create image module
