+++ app/src/main/java/br/com/recipebook/CustomApplication.kt
@@ -2,9 +2,7 @@ package br.com.recipebook
 
 import android.app.Application
 import br.com.recipebook.di.KoinInitializer
-import br.com.recipebook.startup.StartupJob
 import br.com.recipebook.startup.StartupJobsExecutor
-import com.facebook.drawee.backends.pipeline.Fresco
 import io.sentry.android.core.SentryAndroid
 import kotlinx.coroutines.runBlocking
 import org.koin.android.ext.android.inject
@@ -31,11 +29,5 @@ class CustomApplication : Application() {
         }
         // Need to run on main thread
         applicationInitWatcher.watch()
-    }
-}
-
-class ImageLibraryJob(private val application: Application) : StartupJob {
-    override suspend fun invoke() {
-        Fresco.initialize(application)
     }
 }
